Process GTFS data in GeoTrellis and store in DB
The GTFS parser is able to parse all GTFS components, but routes
are the only component that are actually stored in the database
at the moment. Storing the rest of them will be trivial when the
capability exists, and an example is given in GeoTrellisService.scala.

The GTFS parser will soon be published to Maven, but until it is, it
will require a manual step to publish it locally. See the note in
build.sbt for instructions.