Update shape length code to use new interface
The interface of the GTFS editor library has changed in order
to fix a bug with shape lengths. This switches to the new interface,
and updates tests since the length calculation now works.

Asheville tests have also been added.

Note: the GTFS editor does not support the UTM projection by default
as we initially thought it was going to. These calculations are in
WGS84 until we have another task to build in UTM support.