Add Boundary model and Shapefile Celery processing
This adds a model to store the city / region boundary configuration
settings, and a Celery task to construct boundaries from uploaded
shapefiles. The OTIIndicatorsConfig class has been updated to have
pointers to two Boundary objects which will be used as the city and
region boundaries.

This also refactors some of the DataSource code to make it more general,
and puts in place the foundations of a framework for dealing with
Problems associated with a particular DataSource.

Celery tasks have been slightly refactored so that they are still
auto-discoverable, while keeping code specific to different types of
DataSources (e.g. Shapefiles and GTFS feeds) in different files.