Populate memory model from database
Previously, when loading GTFS data, the in-memory model was
created, used for populating the DB, and was cached for later
indicator calculations. This meant that if the GeoTrellis server
was restarted, the indicators would no longer exist.

This change makes it so the GTFS data is loaded and the in-memory
model is discarded. Then when indicators are calculated, they first
pull the data from the database into memory. This is more in-line
with how the service will eventually work -- GTFS and other data are
loaded into the systems, settings are tweaked, and then an indicator
calculation event occurs which will trigger the appropriate data to
be pulled into memory.