Implement city/region boundary uploads
The uploading section of this has been left unabstracted. I attempted
a couple forms of abstraction, but none were satisfactory, so as of
now the upload functionality for the city boundary is a direct copy-
paste-rename of the region boundary upload functionality.

In order to properly abstract this, we will need to create a directive
that is used for all uploads including:
 - GTFS
 - Demographics
 - City boundary
 - Region boundary

There was not enough time to do so during this task, so breaking it out
into a directive and adding tests will need to be a separate task.