Add backend support for importing OSM data
This commit adds necessary models, tasks, and
views to import OpenStreetMap data. The intention
is to kick off the import process once the GTFS
feed is validated and imported.

The angular/front-end code to kick off the job to
import OSM data is going to wait on finishing the
upload polling directive since my changes will be
touching the same code.