Add check for shapes.txt and no shapes_dist_traveled
This adds a warning message to the import process if
both a shapes.txt and no shapes_dist_traveled are available
since length for routes and modes will not be available.

Note: In developing this, no gtfs files with a shape_dist_traveled,
but no shapes.txt file were able to be found.