Fix units and aggregation of length calculator
The units were in meters when kilometers were desired.
Also, the the mode aggregation was using an average when
it should have been using a sum.

Note: these number are now in the ballpark with reality,
but still don't look perfect. SEPTA rail is showing 551
kilometers, where it should be somewhere around 450.
Adding an issue to address this.