Fix missing Indicator uniqueness constraint in migrations
Source: potential bug in django 1.7 migrations,
write up of root issue here:
https://code.djangoproject.com/ticket/23505#ticket

Root issue is that when we drop a column in the migration,
postgres automatically drops the uniqueness constraint, which
django doesn't catch. Thus, when we try to perform migrations
on the uniqueness constraint later, it fails because in all
cases django expects a constraint to exist when it does not.

Had to modify existing migrations, will require a
```
DROP DATABASE transit_indicators;
```
and a reprovision afterwards