Fix simpleMaxBySystem
This would throw an exception if calcByMode returned an empty map.
Haiphong has no weekend routes, so this was occuring when attempting
to calculate the WeeklyServiceHours indicator on a weekend.

Also, max was returning the calculated value of the mode with the
maximum mode id. It has been changed to maxBy(_._2) in order to return
the maximum calculated value.