Save indicator results to DB directly from scala
This commit changes how indicators are save to the
database by saving results directly to the database
via scala. Prior to this commit, results were sent
over the wire to django, which then had to read the
response and then save the results to postgres.

Django could not keep up with saving, serializing,
and deserializing many geometries in quick succession.
This would then lead to timeouts and failures for
indicator processing.

Saving results directly to the database bypasses
this problem.