Push created scenario to scala API
Simplifies new vs existing route logic on route-edit controller
Adds simple client side validation to catch possible common
    errors before a trip.$save is attempted

The isNew flag is added to an OTIRouteModel object
when created via the OTIRouteManager.create() method
so that we know it needs to be POSTed to the API later.

The additional $create endpoint is necessary because of the structure
of the scala API. The normal $save() method cannot be used because it
tries to add the id parameter to the url.