Fix LengthSpec
This isn't so much a fix as changing the expected values to match what
is occurring. As best we can tell, the current test was written at a
point when ScheduledStops were not being filtered out of TransitSystems
if they fell outside of the SamplePeriod for a given indicator
calculation. Manually disabling the stop filtering fixes the LengthSpec,
but is not expected behavior and causes many other tests to fail.

It appears that the LengthSpec was broken by adding / fixing the stops
pruning functionality, but that in all the confusion around the
refactor, it was missed until later when we thought it was due to the
geometry changes. The indicator has been manually verified to be working
correctly using QGIS, so these values should be correct. If they're not,
then it's most likely a problem relating to how transit systems are
constructed, rather than a problem with the Length indicator.