Fix translation loading/persistence
There were a couple different problems here.

* When selecting a new language, the angular translated strings
would change, but the django ones wouldn't. The fix was to do
a basic location.reload instead of a state transition in order
to guarantee everything reloads.

* When refreshing the page, the cookie wasn't correctly being used
to set the language. The final piece of code that was running and
setting the language was only checking for the browser language,
or defaulting to English. The fix was to switch to using the
translateProvider's useCookieStorage setting, which takes will
always give precedence to the language set in the cookie. It also
cleans up the code a bit by moving the cookie setting/getting logic
into its own module.