Remove currentlyCalculating from status checks
Not sure what the purpose of this logic was, but it was causing some problems.
Namely, after indicator calculations were completed, the city wasn't being
added to the list (because currentlyCalculating was set to false), and switching
over to the data page would not show the results of the calculation.

Removing this logic fixes the problems and doesn't seem to cause anything else to
break.