Transform GeoTiff to Lat/Lon before returning
The rasters in the RasterCache are stored in WebMercator, but the
GeoTiff writer only supports EPSG 4326, and does not automatically
transform rasters in other projections. This was causing errors when
trying to use the downloaded GeoTiff under certain circumstances because
its actual coordinates didn't match its stated projection. This
reprojects the raster to EPSG 4326 before serving it.