Add workaround for new sbt-launch location
The sbt-launch jar has moved to bintray. This updates the sbt
bash script to pull from the new location (using wget, since
curl doesn't work on this url for some reason). The newest
version of sbt is not used, since some options have been changed,
which would require other changes in the system.

Also, this sbt file is copied over to the GeoTrellis repo during
provision in order to fix the problem there as well. This is
needed since the GeoTrellis repo is pinned to a certain revision,
and newer versions (that include the sbt fix) will not work
without some code reworking.

In addition, publish-local.sh is no longer being called, because
the gdal artifact it requires is no longer being hosted. Instead,
only the projects we need are individually being published locally.